requirements.txt
langchain>=0.1.0
langchain-community>=0.0.2
langchain-core>=0.1.8
pydantic>=2.4.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
dashscope>=1.13.0
faiss-cpu>=1.7.4
streamlit>=1.28.0
pyyaml>=6.0.1 